Forms, text and images working, time to finish integrating code to main engine. New form Circle added, code moved to engine folder, Draw class for geometric outlines, ScreenFilter and ScreenFade won't work from now, find a way to integrate their functions to the engine later

+++ engine/lib/Tools.hpp
@@ -9,7 +9,7 @@ void Error(std::string);
 int Fps(int);
 double WrapMax(double, double);
 double WrapMinMax(double, double, double);
-float Clamp(float);
+double Clamp(double);
 float SmootherStep(float LowerBound, float HigherBound, float x);
 float SmoothStep (float LowerBound, float HigherBound, float x);
